Enh: Add necessary environment files for Hercules
env/build_hercules_intel.env contains the build environment. Because several of the module versions (and many of the module names) are different from those specified in env/srw_common, srw_common is NOT used in env/build_hercules_intel.env. The build environment relies on the hpc-stack installation in EPIC's project space on /work, and this installation may not be permanent. env/wflow_hercules.env uses my personal Python environment, which was originally created on Orion. This is far from ideal, but it works for now. In the future, we should create a shared Python environment, similar to what is done on Jet/Hera/Orion/etc. Additional changes to regional_workflow to make Hercules a machine option are needed before RRFS can be run on Hercules. The changes here are only enough to compile RRFS succcessfully on Hercules.